Single ShipGroupGenerator added
